requirements.txt
alembic==1.9.4
amqp==5.1.1
Flask
Flask-Admin==1.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
pydantic==1.10.5
SQLAlchemy-i18n==1.1.0
SQLAlchemy-Utils==0.40.0
Werkzeug
mysql-connector-python
flask_sqlalchemy